Use a vector in CACHED instead of an alist
The vector storage is allocated in one go for 20 entries, so it
initially uses more memory. It's more memory efficient though, with a
cross-over point at about 10 entries.

CACHED and (SETF CACHED) are about ~25% faster when MGL-PAX generates
its own READMEs, for example. This flat vector approach is also used
by SBCL for EQ hash tables up to 16 or 32 keys depending on the
platform, which should give some credence to the idea.